Adding system tests for Table Admin API.
We don't explicitly test for

- Table.delete <--> DeleteTable
- Table.list_column_families <--> GetTable

since they are implicitly tested in the other test cases.

We also don't test for

- Table.rename <--> RenameTable
- ColumnFamily.update <--> UpdateColumnFamily

because both of those methods return with the error

> BigtableTableService.{method_name} is not yet implemented

In addition, we noticed that the GC rules in a column family
are not returned when using GetTable.

Also noticed ColumnFamily.__eq__ didn't use gc_rule and that
Table.rename didn't change the locally stored table_id after
success.