Adding temporary skip to test_put_ttl_eviction case.
This is because it has a 3.5 second sleep and we want to avoid
waiting this long while adding more system tests. The skip
will be removed when this file is complete.

Also added a sanity check that the TTL is less than 10 seconds
so we don't change the TTL and cause a 24HR sleep (for example).